# src/Arcus.ClamAV.Api.Client/Generated/swagger.json
{
  "x-generator": "NSwag v14.6.3.0 (NJsonSchema v11.5.2.0 (Newtonsoft.Json v13.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "ClamAv Api",
    "version": "1.0.0"
  },
  "paths": {
    "/healthz": {
      "get": {
        "tags": [
          "Health"
        ],
        "operationId": "HealthCheck",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/version": {
      "get": {
        "tags": [
          "Health"
        ],
        "operationId": "GetVersion",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/scan": {
      "post": {
        "tags": [
          "Scan"
        ],
        "operationId": "ScanSync",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "406": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/scan/async": {
      "post": {
        "tags": [
          "Scan"
        ],
        "operationId": "ScanAsync",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsyncScanResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/scan/async/url": {
      "post": {
        "tags": [
          "Scan"
        ],
        "operationId": "ScanAsyncUrl",
        "requestBody": {
          "x-name": "urlRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanUrlRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsyncScanResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/scan/json": {
      "post": {
        "tags": [
          "Scan"
        ],
        "operationId": "ScanJson",
        "requestBody": {
          "x-name": "jsonPayload",
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonScanResult"
                }
              }
            }
          },
          "406": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonScanResult"
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/scan/async/{jobId}": {
      "get": {
        "tags": [
          "Scan"
        ],
        "operationId": "GetScanStatus",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/scan/jobs": {
      "get": {
        "tags": [
          "Scan"
        ],
        "operationId": "ListJobs",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobsListResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
      "VersionResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "clamAvVersion": {
            "type": "string"
          }
        }
      },
      "ScanResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "status": {
            "type": "string"
          },
          "engine": {
            "type": "string"
          },
          "malware": {
            "type": "string",
            "nullable": true
          },
          "fileName": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "scanDurationMs": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "AsyncScanResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "jobId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string"
          },
          "statusUrl": {
            "type": "string"
          },
          "sourceUrl": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ScanUrlRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "url": {
            "type": "string"
          },
          "isBase64": {
            "type": "boolean"
          }
        }
      },
      "JsonScanResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "status": {
            "type": "string"
          },
          "itemsScanned": {
            "type": "integer",
            "format": "int32"
          },
          "base64ItemsFound": {
            "type": "integer",
            "format": "int32"
          },
          "malware": {
            "type": "string",
            "nullable": true
          },
          "infectedItem": {
            "type": "string",
            "nullable": true
          },
          "scanDurationMs": {
            "type": "number",
            "format": "double"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScannedItemDetail"
            }
          }
        }
      },
      "ScannedItemDetail": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string"
          },
          "malware": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ScanStatusResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "jobId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "fileSize": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "engine": {
            "type": "string",
            "nullable": true
          },
          "malware": {
            "type": "string",
            "nullable": true
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "scanDurationMs": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        }
      },
      "JobsListResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "jobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobSummary"
            }
          },
          "count": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "JobSummary": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "jobId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "fileSize": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "scanDurationMs": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        }
      }
    }
  }
}